pyproject.toml
[project]
name = "cnn-cat-vs-dog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.1.1",
    "streamlit>=1.54.0",
    "tensorflow>=2.20.0",
]
